context menu support
This fix the behavior on the right click -> open in new tab action.

On the pad content the value of the href is changed everytime there is a right
click, as any value we set there gets overwritten by etherpad. On the chats the
situation is different, the href value is permanent so we add a class to know that
we changed the value of the href attribute and not add "../redirect#" again.

These changes are needed for Firefox ESR and Tor Browser at least.