Remove the chance to click to execute javascript code
This removes the option for an user to click and execute javascript when a
possible XSS is detected. The concept of XSS is not universally known, meaning
that is still possible for somebody to click an address as `javascript:(alert)`
and run the javascript in their browse.

If the user still wants to load the content they should manually copy / paste it
in their browser's address bar.